// src/level/LDTK.d.ts
export interface TilesetRectangle {
	/**
	 * Height in pixels
	 */
	h: number;
	/**
	 * UID of the tileset
	 */
	tilesetUid: number;
	/**
	 * Width in pixels
	 */
	w: number;
	/**
	 * X pixels coordinate of the top-left corner in the Tileset image
	 */
	x: number;
	/**
	 * Y pixels coordinate of the top-left corner in the Tileset image
	 */
	y: number;
}
// ...
export interface EntityInstance {
	/**
	 * Grid-based coordinates (`[x,y]` format)
	 */
	__grid: number[];
	/**
	 * Entity definition identifier
	 */
	__identifier: string;
	/**
	 * Pivot coordinates  (`[x,y]` format, values are from 0 to 1) of the Entity
	 */
	__pivot: number[];
	/**
	 * The entity "smart" color, guessed from either Entity definition, or one its field
	 * instances.
	 */
	__smartColor: string;
	/**
	 * Array of tags defined in this Entity definition
	 */
	__tags: string[];
	/**
	 * Optional TilesetRect used to display this entity (it could either be the default Entity
	 * tile, or some tile provided by a field value, like an Enum).
	 */
	__tile?: TilesetRectangle | null;
	/**
	 * Reference of the **Entity definition** UID
	 */
	defUid: number;
	/**
	 * An array of all custom fields and their values.
	 */
	fieldInstances: FieldInstance[];
	/**
	 * Entity height in pixels. For non-resizable entities, it will be the same as Entity
	 * definition.
	 */
	height: number;
	/**
	 * Unique instance identifier
	 */
	iid: string;
	/**
	 * Pixel coordinates (`[x,y]` format) in current level coordinate space. Don't forget
	 * optional layer offsets, if they exist!
	 */
	px: number[];
	/**
	 * Entity width in pixels. For non-resizable entities, it will be the same as Entity
	 * definition.
	 */
	width: number;
}

export interface FieldInstance {
	/**
	 * Field definition identifier
	 */
	__identifier: string;
	/**
	 * Optional TilesetRect used to display this field (this can be the field own Tile, or some
	 * other Tile guessed from the value, like an Enum).
	 */
	__tile?: TilesetRectangle | null;
	/**
	 * Type of the field, such as `Int`, `Float`, `String`, `Enum(my_enum_name)`, `Bool`,
	 * etc.<br/>  NOTE: if you enable the advanced option **Use Multilines type**, you will have
	 * "*Multilines*" instead of "*String*" when relevant.
	 */
	__type: string;
	/**
	 * Actual value of the field instance. The value type varies, depending on `__type`:<br/>
	 * - For **classic types** (ie. Integer, Float, Boolean, String, Text and FilePath), you
	 * just get the actual value with the expected type.<br/>   - For **Color**, the value is an
	 * hexadecimal string using "#rrggbb" format.<br/>   - For **Enum**, the value is a String
	 * representing the selected enum value.<br/>   - For **Point**, the value is a
	 * [GridPoint](#ldtk-GridPoint) object.<br/>   - For **Tile**, the value is a
	 * [TilesetRect](#ldtk-TilesetRect) object.<br/>   - For **EntityRef**, the value is an
	 * [EntityReferenceInfos](#ldtk-EntityReferenceInfos) object.<br/><br/>  If the field is an
	 * array, then this `__value` will also be a JSON array.
	 */
	__value: any;
	/**
	 * Reference of the **Field definition** UID
	 */
	defUid: number;
	/**
	 * Editor internal raw values
	 */
	realEditorValues: any[];
}
// ...
export interface IntGridValueInstance {
	/**
	 * Coordinate ID in the layer grid
	 */
	coordId: number;
	/**
	 * IntGrid value
	 */
	v: number;
}

export interface LayerInstance {
	/**
	 * Grid-based height
	 */
	__cHei: number;
	/**
	 * Grid-based width
	 */
	__cWid: number;
	/**
	 * Grid size
	 */
	__gridSize: number;
	/**
	 * Layer definition identifier
	 */
	__identifier: string;
	/**
	 * Layer opacity as Float [0-1]
	 */
	__opacity: number;
	/**
	 * Total layer X pixel offset, including both instance and definition offsets.
	 */
	__pxTotalOffsetX: number;
	/**
	 * Total layer Y pixel offset, including both instance and definition offsets.
	 */
	__pxTotalOffsetY: number;
	/**
	 * The definition UID of corresponding Tileset, if any.
	 */
	__tilesetDefUid?: number | null;
	/**
	 * The relative path to corresponding Tileset, if any.
	 */
	__tilesetRelPath?: null | string;
	/**
	 * Layer type (possible values: IntGrid, Entities, Tiles or AutoLayer)
	 */
	__type: string;
	/**
	 * An array containing all tiles generated by Auto-layer rules. The array is already sorted
	 * in display order (ie. 1st tile is beneath 2nd, which is beneath 3rd etc.).<br/><br/>
	 * Note: if multiple tiles are stacked in the same cell as the result of different rules,
	 * all tiles behind opaque ones will be discarded.
	 */
	autoLayerTiles: TileInstance[];
	entityInstances: EntityInstance[];
	gridTiles: TileInstance[];
	/**
	 * Unique layer instance identifier
	 */
	iid: string;
	/**
	 * **WARNING**: this deprecated value is no longer exported since version 1.0.0  Replaced
	 * by: `intGridCsv`
	 */
	intGrid?: IntGridValueInstance[] | null;
	/**
	 * A list of all values in the IntGrid layer, stored in CSV format (Comma Separated
	 * Values).<br/>  Order is from left to right, and top to bottom (ie. first row from left to
	 * right, followed by second row, etc).<br/>  `0` means "empty cell" and IntGrid values
	 * start at 1.<br/>  The array size is `__cWid` x `__cHei` cells.
	 */
	intGridCsv: number[];
	/**
	 * Reference the Layer definition UID
	 */
	layerDefUid: number;
	/**
	 * Reference to the UID of the level containing this layer instance
	 */
	levelId: number;
	/**
	 * An Array containing the UIDs of optional rules that were enabled in this specific layer
	 * instance.
	 */
	optionalRules: number[];
	/**
	 * This layer can use another tileset by overriding the tileset UID here.
	 */
	overrideTilesetUid?: number | null;
	/**
	 * X offset in pixels to render this layer, usually 0 (IMPORTANT: this should be added to
	 * the `LayerDef` optional offset, so you should probably prefer using `__pxTotalOffsetX`
	 * which contains the total offset value)
	 */
	pxOffsetX: number;
	/**
	 * Y offset in pixels to render this layer, usually 0 (IMPORTANT: this should be added to
	 * the `LayerDef` optional offset, so you should probably prefer using `__pxTotalOffsetX`
	 * which contains the total offset value)
	 */
	pxOffsetY: number;
	/**
	 * Random seed used for Auto-Layers rendering
	 */
	seed: number;
	/**
	 * Layer instance visibility
	 */
	visible: boolean;
}

/**
 * This structure represents a single tile from a given Tileset.
 */
export interface TileInstance {
	/**
	 * Internal data used by the editor.<br/>  For auto-layer tiles: `[ruleId, coordId]`.<br/>
	 * For tile-layer tiles: `[coordId]`.
	 */
	d: number[];
	/**
	 * "Flip bits", a 2-bits integer to represent the mirror transformations of the tile.<br/>
	 * - Bit 0 = X flip<br/>   - Bit 1 = Y flip<br/>   Examples: f=0 (no flip), f=1 (X flip
	 * only), f=2 (Y flip only), f=3 (both flips)
	 */
	f: number;
	/**
	 * Pixel coordinates of the tile in the **layer** (`[x,y]` format). Don't forget optional
	 * layer offsets, if they exist!
	 */
	px: number[];
	/**
	 * Pixel coordinates of the tile in the **tileset** (`[x,y]` format)
	 */
	src: number[];
	/**
	 * The *Tile ID* in the corresponding tileset.
	 */
	t: number;
}